msm8953-common: Decrease clockspeed and duration for app launch hint
Since the lineage qcom powerhal now sends POWER_HINT_LAUNCH when it
detects a long interaction, it now means that there are going to be more
times when the CPU cores would be pinned to the max speed for at least 3
seconds, which is detrimental to battery life.

This commit lowers the clock speed and duration for the app launch power
hint to be 1689MHz and 2 seconds.

Co-Authored-By: Stefan <stefan.zivkovic995@gmail.com>